requirements.txt
# ===========================================
# K Stock Trader - Consolidated Requirements
# ===========================================

# Core
pandas>=2.0.0
requests>=2.31.0
loguru>=0.7.0
pyyaml>=6.0
websockets>=11.0
aiohttp>=3.9.0

# OMS Server
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
asyncpg>=0.29.0

# PCIM: YouTube + Gemini
feedparser>=6.0.0
yt-dlp>=2024.1.0
google-genai>=0.5.0

# Timezone
pytz>=2024.1

# Optional: CLI
rich>=13.0.0
